+++ cryptol-remote-api/python/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Revision history for `cryptol` Python package
 
+## 2.11.6 -- 2021-09-10
+
+* Add a `timeout` field to the `CryptolConnection` class which can be used
+  to set a default timeout for all RPC interactions.
+* Add an optional `timeout` keyword parameter to each `CryptolConnection` method
+  which can specify a timeout for a particular method call.
+* Add an `interrupt` method to the `CryptolConnection` class which interrupts
+  any active requests on the server.
+
 ## 2.11.5 -- 2021-08-25
 
 * From argo: Change the behavior of the `Command` `state` method so that after